requirements.txt
annotated-doc
annotated-types
anyio
click
colorama
fastapi
greenlet
h11
idna
psycopg2
pydantic
pydantic_core
PyMySQL
python-dotenv
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn